charmap, default latin transliteration, tests not included, tests corrected
Asciize function now uses charmap - a dictionary mapping characters to their
transliterations. Once new character (non ascii) in the source string is found,
it is added to charmap; before using unicode database, current character is
checked to be present in the charmap.

If asciize is called with no charmap, by default
unicode2ascii.latin_transliteration will be used - the effect of asciizing
latin text will be quite "better" (eg. Æ will be asciized to ae).

Tests are changed to meet the presence of charmap - in general there are less
'_' in expected strings. Tests are not a part of the package now - they have
form of ready-to-build separate package.